MAINT: Add user/role page

+++ core/src/main/java/org/trade/core/persistent/domain/DomainServiceImpl.java
@@ -2,8 +2,6 @@ package org.trade.core.persistent.domain;
 
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-
 /**
  * @author Simon Allen
  * @version $Revision: 1.0 $
@@ -18,9 +16,9 @@ public class DomainServiceImpl implements DomainService {
         this.domainRepository = domainRepository;
     }
 
-    public Optional<Domain> findDomainByName(String name) {
+    public Domain findDomainByName(String name) {
 
-        return domainRepository.findByName(name);
+        return domainRepository.findByName(name).orElse(null);
     }
 
     public Domain saveDomain(Domain domain) {
@@ -29,6 +27,11 @@ public class DomainServiceImpl implements DomainService {
     }
 
     public void deleteDomain(Domain domain) {
+
+        if (null == domain) {
+
+            return;
+        }
 
         domainRepository.delete(domain);
     }
